jkkljkjlkjllvhvgvvh IM SICKKKK OF THIS!!!!!!!!!!!!!!

+++ Codes/Cleaned SA Code/inputs.py
@@ -64,7 +64,7 @@ class variables:
       self.P0      = np.zeros(variables.NK)
       self.P1      = np.zeros(variables.NK, dtype=np.float64)
       self.P11     = np.zeros(variables.NK)
-      self.PB      = np.zeros(variables.NK)
+      self.PB      = np.zeros(variables.NK, dtype=np.float64)
       self.PIN     = 12262500
       self.PM0     = np.zeros(variables.NK)
